academic_research_assistant/requirements.txt
streamlit==1.32.0
fastapi==0.109.0
uvicorn==0.27.0
neo4j==5.24.2
transformers==4.37.0
torch==2.2.0
scholarly==1.7.11
python-dotenv==1.0.0
outlines==0.0.30
pydantic==2.6.1
arxiv==2.0.0